+++ src/HaoRan/ImageFilter/ColorQuantizeFilter.java
@@ -37,9 +37,9 @@ public class ColorQuantizeFilter implements IImageFilter{
                  float quanR = (((float) ((int) (r * 0.003921569f * levels))) / levels) * 255f;
                  float quanG = (((float) ((int) (g * 0.003921569f * levels))) / levels) * 255f;
                  float quanB = (((float) ((int) (b * 0.003921569f * levels))) / levels) * 255f;
-                 r = (quanR > 255f) ? 255 : ((quanR < 0f) ? 0 : ((byte) quanR));
-                 g = (quanG > 255f) ? 255 : ((quanG < 0f) ? 0 : ((byte) quanG));
-                 b = (quanB > 255f) ? 255 : ((quanB < 0f) ? 0: ((byte) quanB));
+                 r = (quanR > 255f) ? 255 : ((quanR < 0f) ? 0 : ((int)quanR));
+                 g = (quanG > 255f) ? 255 : ((quanG < 0f) ? 0 : ((int) quanG));
+                 b = (quanB > 255f) ? 255 : ((quanB < 0f) ? 0: ((int) quanB));
                  imageIn.setPixelColor(x,y,r,g,b);
              }
         }
